fix: useEnhancer and useHistory
